requirements.txt
tensorflow>=1.3.0
opencv-python>=3
Pillow>=4.3.0
